ENH organize benchmark functions
